Create README.md
Abstract
This application is built for Android devices. It is used to locate places for buying essential items and can purchase from doorsteps or curbside pickup. It is based on a token system that means it will generate tokens for the customers and an estimated wait time for them. It is easy to use as no technical knowledge is required to use this application. This application will fully support any android device.
In this application, we used Java, Android for the development of the application and Extensible Markup Language (XML) and Material Design for Application designing and firebase for the database.
This application is developed to Assist the public so that during this pandemic they can properly and safely tackle the situations. 
To develop an application that can be used for the betterment of people during these difficult times so that during pandemics everyone will be able to be safe and protected.
Introduction
Long queues outside stores are very common now. We don’t have to stay in time-consuming and energy-draining queues anymore. This idea will help everyone who goes shopping and to get any kind of services and those who don’t like to wait in long queues. Long queues to get any service due to COVID 19. By implementing our idea people can save time and energy. It is difficult for people to stay outside for long in challenging weather. Some people might want to buy just one or two items. It is really frustrating to stay in long queues for that. For example, grocery shopping, which happens frequently. It’s frustrating and time-consuming to stay in long queues. The customer can feel that even going grocery shopping is a burden to themselves. They could get a feeling that they could have used that time for something productive. They don’t have to stand in queue for long and will not waste customer time. It brings back the convenience is the shopping. Also, we are implementing physical distancing also by making people not stay around.

 
Features
Effective queue management: 
The queue will be managed properly to avoid misunderstanding and chaos, as well as overcrowding. 
Providing customer support:
Customer support and assistance will be provided to the customers as per their needs.
Adding features and updating the application as per the requirement:
Implementation and updating of the Application will be done on a regular basis to provide a better user experience.
Shops 
Revenue will be generated from the shop maintaining teams as well as the organizing committee of an event who wants to manage the public and queues.
Future plans:
In-App purchase: 
in-app purchases are the purchases the users can make to make changes to the application as such if a user doesn’t want to see the annoying ads and pop-ups so he can upgrade the services of the application by paying for the application after which ads will be completely removed from the application.
Advertisements: 
Ads will be shown in the application as pop-ups and banners so the revenues can be generated from the businesses and the organizers who want to publish their ads on our application.